refactor: merge shell tests
